requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
numpy==1.16.0
opencv-python==4.0.0.21
Pillow==5.4.1
requests==2.21.0
urllib3==1.24.1
